Consolidate/move fileclient/fileserver tests to conform to naming convention
